+++ src/components/Gifs/Gif.js
@@ -1,0 +1,3 @@
+export default function Gif(url) {
+  return `<img class="m-1" src="${url}" />`;
+}
