+++ src/data/dataStore.js
@@ -1,0 +1,16 @@
+const dataStore = {
+  cache: {
+    gifs: null,
+  },
+  uiState: {
+    /*
+     * processing
+     * ready
+     *
+     * */
+    state: null,
+    error: null,
+  },
+};
+
+export default dataStore;
