--- src/js/app.js
@@ -1,5 +1,0 @@
-/* eslint-disable no-console */
-
-export default () => {
-  console.log('starting app');
-};
